Populate ok

+++ Labo.ASP/Models/AdressModel.cs
@@ -12,6 +12,6 @@ namespace Labo.ASP.Models
         public string NumRue { get; set; }
         public string Ville { get; set; }
         public string Pays { get; set; }
-        public int CodePostal { get; set; }
+        public long CodePostal { get; set; }
     }
 }
